api/requirements.txt
django==4.2
djangorestframework==3.14.0
psycopg2-binary==2.9.9
black==24.2.0
confluent-kafka==2.3.0
